refactor: split types and validations in new file
