+++ app/dashboard/campaigns/create/page.tsx
@@ -1,8 +1,10 @@
 "use client";
 
 import { useRouter } from "next/navigation";
-import { CampaignForm, CampaignFormValues } from "@/components/campaigns/CampaignForm";
+import { CampaignForm } from "@/components/campaigns/CampaignForm";
 import { copyText as t } from "@/components/campaigns/copy";
+import { CampaignFormValues } from "@/components/campaigns/types";
+import { createCampaignMockAction } from "@/lib/actions/campaignActions";
 
 export default function CampaignCreatePage() {
   const router = useRouter();
@@ -10,6 +12,7 @@ export default function CampaignCreatePage() {
   const handleSubmit = async (data: CampaignFormValues) => {
     console.log("Form Submitted:", data);
     // TODO: Implement campaign creation
+    await createCampaignMockAction(data);
   };
 
   const handleCancel = () => {
@@ -21,7 +24,7 @@ export default function CampaignCreatePage() {
       <div className="flex items-center justify-between">
         <h1 className="text-2xl font-bold tracking-tight">{t.pageTitle}</h1>
       </div>
-      <CampaignForm 
+      <CampaignForm
         onSubmit={handleSubmit}
         onCancel={handleCancel}
       />
